+++ scripts/generate-types.js
@@ -62,21 +62,24 @@ const generateTypes = (definitions) => {
 import type { MediaType } from 'openapi-typescript-helpers';import { paths } from '../api';
 import { WalletService } from '../services/wallet-service';
 \n\n`,
-    `export type Environment = 'dev' | 'prod';
+    `export type Environment = 'dev' | 'prod';export type TokenManagementMode = 'jwt' | 'privateKey';
 \n`,
     `export interface Config {
-  privateKey: string;
+    jwt?:string
   environment: Environment;
   logging?: boolean;
-  walletType: WalletTypeEnum;
+  wallet?: {
+    privateKey: string;
+    walletType: WalletTypeEnum;
+  };
 }\n\n`,
     `export interface WalletSignMessageType {
   signature: string;
   signingKey: string;
 }\n\n`,
     `export enum WalletTypeEnum {
-  ED25519 = 'ED25519',
-  SECP256K1 = 'SECP256K1',
+  Etherum = 'etherum',
+  Solana = 'solana',
 }\n\n`,
     `export interface JWTData {
   did: string;