+++ src/helpers/helper.ts
@@ -4,6 +4,7 @@ import {
   CustomConfig,
   CustomMiddlewareWithVariables,
   Environment,
+  TokenManagementMode,
 } from '../common/types';
 import { WalletService } from '../services/wallet-service';
 import { MediaType } from 'openapi-typescript-helpers';
@@ -23,18 +24,44 @@ import { routes } from '../common/routes';
  * `environment` is not `'dev'`, it throws an error with the message 'No valid url found!. Use sandbox
  * or production url'.
  */
-export const parameterChecker = (environment: Environment): string => {
+export const parameterChecker = (
+  environment: Environment,
+  jwt?: string,
+  privateKey?: string,
+): { url: string; mode: TokenManagementMode; value: string } => {
+  let mode: TokenManagementMode;
+  let value = '';
   if (!environment)
     throw new Error('No url found!.Use either sandbox or production env');
 
-  const urls = ['https://dev.api.gateway.tech'];
+  if (privateKey) {
+    mode = 'privateKey';
+    value = privateKey;
+  } else if (jwt) {
+    mode = 'jwt';
+    value = jwt;
+    if (!checkJWTTokenExpiration(jwt))
+      throw new Error('The provided token is expired or invalid.');
+  } else {
+    throw new Error('Need jwt or private key');
+  }
 
-  if (environment === 'dev') return urls[0];
+  const urls = ['https://dev.api.gateway.tech', 'https://api.gateway.tech'];
+
+  if (environment === 'dev') return { url: urls[0], mode, value };
+  else if (environment === 'prod') return { url: urls[1], mode, value };
   else throw new Error('No valid url found!. Use sandbox or production url');
 };
 
 let accessToken: string | undefined = undefined;
 
+/**
+ * The function `checkJWTTokenExpiration` verifies if a JWT token is expired or not.
+ * @param {string} existinToken - The `existinToken` parameter is a string representing a JWT token
+ * that you want to check for expiration.
+ * @returns The function `checkJWTTokenExpiration` returns a boolean value. It returns `true` if the
+ * JWT token is valid and has not expired, and `false` if the token is invalid or has expired.
+ */
 export const checkJWTTokenExpiration = (existinToken: string): boolean => {
   try {
     const decodedToken = jwt.decode(existinToken) as JWTData | null;
@@ -60,7 +87,7 @@ export const issueJWT = async (
 ) => {
   const auth = new Auth(client);
 
-  const message = await auth.generateSignMessage();
+  const message = await auth.getMessage();
   const signatureDetails = await wallet.signMessage(message);
 
   const jwt = await auth.login({